requirements.txt
flask==2.3.3
python-binance==1.0.19
flask-socketio==5.3.6
python-socketio==5.8.0
python-engineio==4.7.1
requests==2.31.0
cryptography==41.0.4
python-dotenv==1.0.0
pyinstaller
